+++ packages/frontend/src/ee/features/aiCopilot/components/Admin/AgentsFilter.tsx
@@ -190,7 +190,6 @@ const AgentsFilter: FC<AgentsFilterProps> = ({
                                                                         size={
                                                                             16
                                                                         }
-                                                                        variant="filled"
                                                                         name={
                                                                             agent.name
                                                                         }